+++ src/components/display/ToggleDisplay.jsx
