+++ src/components/index.jsx
@@ -1,7 +1,7 @@
-export * from "./Routers";
+export * from "./router/Routers";
 export * from "./AppLayout";
-export * from "./MainNav";
-export * from "./ToggleDisplay";
+export * from "./nav/MainNav";
+export * from "./display/ToggleDisplay";
 export * from "./common/Button";
 export * from "./common/Nav";
 export * from "./wallet/CoinContainer";
